Replace Guava RateLimiter with a bursty token bucket. (#158)

+++ agent/src/test/java/com/spotify/ffwd/FfwdConfigurationTest.java
@@ -93,7 +93,7 @@ public class FfwdConfigurationTest {
         CoreOutputManager outputManager = getOutputManager(null);
 
         assertEquals(100, outputManager.getTtl());
-        assertEquals(Double.valueOf(1000), outputManager.getRateLimit());
+        assertEquals(Long.valueOf(1000), outputManager.getRateLimit());
     }
 
     @Test
